Add an MJPEG test stream for USB camera.
The camera claims to support 30fps 1920x1080 MJPEG but in fact
delivers approx 8fps at this res.

nvjpegdec complains about the jpeg stream but the uvc demux helps.

nvjpegdec claims to provide NVMM or raw video out but only seems
to work with raw video, though not direct to the java appsink
for some reason. Back to back nvvidconvs help with this.

This pipeline is a mess, and the camera sucks.